Move settings button within Header component
